// src/app/models/Frequentadores.ts
export class Frequentadores {
  frequentadorId: number;
  nome: string;
  cpf: string;
  dataNascimento: Date;
  sexo: string;
  telefone: string;
  email: string;
  imagem: string;
  bloqueadoSite: boolean;
  bloqueado: boolean;
  bloqueio: string;
  ativo: boolean;
  tipoPessoaId: string;
  tipoPessoa: string;
}